Cambios relativos a Nueva entidad padre Usuario
Se ha añadido una clase padre Usuario de la que heredan  Admin,Client,Reporter y Company

+++ TFG/src/main/java/com/project/controllers/FactureClientController.java
@@ -30,9 +30,9 @@ public class FactureClientController {
 	// -------------------------- List Facture by Client ----------------------------------
 	@CrossOrigin
 	@RequestMapping(value = "/myFactures/{clientId}", method = RequestMethod.GET)
-	public List<Facture> listFactureByClient(@PathVariable Long clientId) {
+	public List<Facture> listFactureByClient(@PathVariable int clientId) {
 		return factureService.findFactureByClient(clientId);
-	}
+	}  
 
 	// ---------------------------- Show facture----------------------------------------------------------
 	@CrossOrigin
