+++ TFG/src/main/java/com/project/repositories/FactureRepository.java
@@ -12,9 +12,9 @@ import com.project.domain.Facture;
 public interface FactureRepository extends JpaRepository<Facture, Long> {
 	
 	@Query("select f from Facture f where f.client.id = ?1")
-	List<Facture> findFacturesByClient(Long clientId);
+	List<Facture> findFacturesByClient(int clientId);
 	
 	@Query("select f From Facture f where f.company.id =?1")
-	List<Facture> findFactureByCompany(Long companyId);
+	List<Facture> findFactureByCompany(int companyId);
 
 }
